organized SASS/CSS files structure
organized sass files and css files into separate folders for organization purposes.

PLEASE USE THE TESTING COMPONENT AS AN EXAMPLE OF HOW TO USE SASS IN your COMPONENT file.
