+++ components/sections/Testimonials.jsx
@@ -1,6 +1,7 @@
 import React, { Component } from "react"
 import "react-responsive-carousel/lib/styles/carousel.min.css"
 import { Carousel } from "react-responsive-carousel"
+
 export default class Testimonials extends Component {
   render() {
     return (
