--- styles/sass/index.module.scss
@@ -1,5 +1,0 @@
-.testsass{
-    background-color: rgb(250, 3, 45) !important;
-    color: rgb(255, 255, 255) !important; 
-    text-align: center;
-}
